apps/ml/requirements.txt
# ML & Forecasting (versões pré-compiladas para Windows)
prophet>=1.1
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3

# API Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1

# Database
supabase==2.3.4
python-dotenv==1.0.1

# Utils
python-dateutil==2.8.2
pytz==2024.1
